+++ flowman-core/src/test/scala/com/dimajix/flowman/execution/JobCoordinatorTest.scala
@@ -50,7 +50,7 @@ class JobCoordinatorTest extends AnyFlatSpec with Matchers with MockFactory with
             targets = Seq(TargetIdentifier("some_target"))
         )
 
-        (targetP.instantiate _).expects(*,None).returns(target)
+        (targetP.instantiate _).expects(*, None).returns(target)
         (target.name _).expects().atLeastOnce().returns("some_target")
         (target.identifier _).expects().atLeastOnce().returns(TargetIdentifier("some_target"))
         (target.before _).expects().atLeastOnce().returns(Seq())
@@ -69,7 +69,7 @@ class JobCoordinatorTest extends AnyFlatSpec with Matchers with MockFactory with
         (target.execute _).expects(*, Phase.BUILD).returning(TargetResult(target, Phase.BUILD, Status.SUCCESS, Instant.now()))
 
         val coord = new JobCoordinator(session)
-        coord.execute(job, Lifecycle.ofPhase(Phase.BUILD)) should be (Status.SUCCESS)
+        coord.execute(job, Lifecycle.ofPhase(Phase.BUILD)) should be(Status.SUCCESS)
     }
 
     it should "support parallel execution" in {
@@ -87,11 +87,11 @@ class JobCoordinatorTest extends AnyFlatSpec with Matchers with MockFactory with
             parameters = Seq(
                 Job.Parameter("pdate", DateType)
             ),
-            phases = Map(
-                Phase.VALIDATE -> PhaseExecutionPolicy.FIRST,
-                Phase.CREATE -> PhaseExecutionPolicy.NEVER,
-                Phase.BUILD -> PhaseExecutionPolicy.ALWAYS,
-                Phase.VERIFY -> PhaseExecutionPolicy.LAST
+            executions = Seq(
+                Job.Execution(Phase.VALIDATE, PhaseExecutionPolicy.FIRST),
+                Job.Execution(Phase.CREATE, PhaseExecutionPolicy.NEVER),
+                Job.Execution(Phase.BUILD, PhaseExecutionPolicy.ALWAYS),
+                Job.Execution(Phase.VERIFY, PhaseExecutionPolicy.LAST)
             )
         )
 
@@ -117,7 +117,7 @@ class JobCoordinatorTest extends AnyFlatSpec with Matchers with MockFactory with
         coord.execute(job, Lifecycle.ofPhase(Phase.VERIFY), job.parseArguments(Map("pdate:start" -> "2022-11-01", "pdate:end" -> "2022-11-09"))) should be(Status.SUCCESS)
     }
 
-    it should "use explicit phase configuration" in {
+    it should "support different phase execution policies" in {
         val firstTarget = mock[Target]
         val midTarget = mock[Target]
         val lastTarget = mock[Target]
@@ -134,15 +134,15 @@ class JobCoordinatorTest extends AnyFlatSpec with Matchers with MockFactory with
             parameters = Seq(
                 Job.Parameter("pdate", DateType)
             ),
-            phases = Map(
-                Phase.VALIDATE -> PhaseExecutionPolicy.FIRST,
-                Phase.CREATE -> PhaseExecutionPolicy.NEVER,
-                Phase.BUILD -> PhaseExecutionPolicy.ALWAYS,
-                Phase.VERIFY -> PhaseExecutionPolicy.LAST
+            executions = Seq(
+                Job.Execution(Phase.VALIDATE, PhaseExecutionPolicy.FIRST),
+                Job.Execution(Phase.CREATE, PhaseExecutionPolicy.NEVER),
+                Job.Execution(Phase.BUILD, PhaseExecutionPolicy.ALWAYS),
+                Job.Execution(Phase.VERIFY, PhaseExecutionPolicy.LAST)
             )
         )
 
-        (targetP.instantiate _).expects(*,*).atLeastTwice().onCall { (ctx, _) =>
+        (targetP.instantiate _).expects(*, *).atLeastTwice().onCall { (ctx, _) =>
             val pdate = ctx.environment.get("pdate").get
             if (pdate == DateType.parse("2022-11-01"))
                 firstTarget
@@ -161,7 +161,7 @@ class JobCoordinatorTest extends AnyFlatSpec with Matchers with MockFactory with
         (firstTarget.provides _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
         (firstTarget.digest _).expects(Phase.VALIDATE).atLeastOnce().returns(TargetDigest("default", "project", "some_target", Phase.VALIDATE))
         (firstTarget.dirty _).expects(*, Phase.VALIDATE).returns(Yes)
-        (firstTarget.execute _).expects(*, Phase.VALIDATE) returning(TargetResult(firstTarget, Phase.VALIDATE, Status.SUCCESS, Instant.now()))
+        (firstTarget.execute _).expects(*, Phase.VALIDATE) returning (TargetResult(firstTarget, Phase.VALIDATE, Status.SUCCESS, Instant.now()))
         (firstTarget.requires _).expects(Phase.BUILD).atLeastOnce().returns(Set())
         (firstTarget.provides _).expects(Phase.BUILD).atLeastOnce().returns(Set())
         (firstTarget.digest _).expects(Phase.BUILD).atLeastOnce().returns(TargetDigest("default", "project", "some_target", Phase.BUILD))
@@ -199,5 +199,169 @@ class JobCoordinatorTest extends AnyFlatSpec with Matchers with MockFactory with
 
         val coord = new JobCoordinator(session)
         coord.execute(job, Lifecycle.ofPhase(Phase.VERIFY), job.parseArguments(Map("pdate:start" -> "2022-11-01", "pdate:end" -> "2022-11-11"))) should be(Status.SUCCESS)
+    }
+
+    it should "support target lists per execution phase" in {
+        val target1 = mock[Target]
+        val target2 = mock[Target]
+        val target3 = mock[Target]
+        val targetP1 = mock[Prototype[Target]]
+        val targetP2 = mock[Prototype[Target]]
+        val targetP3 = mock[Prototype[Target]]
+        val project = Project(
+            "project",
+            targets = Map(
+                "target1" -> targetP1,
+                "target2" -> targetP2,
+                "target3" -> targetP3
+            )
+        )
+        val session = Session.builder().withSparkSession(spark).build()
+        val context = session.getContext(project)
+        val job = Job(
+            Job.Properties(context),
+            targets = Seq(
+                TargetIdentifier("target1"),
+                TargetIdentifier("target2"),
+                TargetIdentifier("target3")
+            ),
+            executions = Seq(
+                Job.Execution(Phase.VALIDATE, targets=Seq(".*".r)),
+                Job.Execution(Phase.CREATE, targets=Seq(".*1".r)),
+                Job.Execution(Phase.BUILD, targets=Seq(".*x".r))
+            )
+        )
+
+        (targetP1.instantiate _).expects(*, *).once().returns(target1)
+        (target1.name _).expects().atLeastOnce().returns("target1")
+        (target1.identifier _).expects().atLeastOnce().returns(TargetIdentifier("target1"))
+        (target1.before _).expects().atLeastOnce().returns(Seq())
+        (target1.after _).expects().atLeastOnce().returns(Seq())
+        (target1.phases _).expects().atLeastOnce().returns(Set(Phase.VALIDATE, Phase.CREATE, Phase.BUILD, Phase.VERIFY))
+        (target1.metadata _).expects().atLeastOnce().returns(Metadata(name = "target1", kind = "target", category = "target"))
+        (target1.requires _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target1.provides _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target1.digest _).expects(Phase.VALIDATE).atLeastOnce().returns(TargetDigest("default", "project", "target1", Phase.VALIDATE))
+        (target1.dirty _).expects(*, Phase.VALIDATE).returns(Yes)
+        (target1.execute _).expects(*, Phase.VALIDATE) returning (TargetResult(target1, Phase.VALIDATE, Status.SUCCESS, Instant.now()))
+        (target1.requires _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target1.provides _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target1.digest _).expects(Phase.CREATE).atLeastOnce().returns(TargetDigest("default", "project", "target1", Phase.BUILD))
+        (target1.dirty _).expects(*, Phase.CREATE).returns(Yes)
+        (target1.execute _).expects(*, Phase.CREATE).returning(TargetResult(target1, Phase.CREATE, Status.SUCCESS, Instant.now()))
+        (target1.requires _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+        (target1.provides _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+        (target1.digest _).expects(Phase.VERIFY).atLeastOnce().returns(TargetDigest("default", "project", "target1", Phase.VERIFY))
+        (target1.dirty _).expects(*, Phase.VERIFY).returns(Yes)
+        (target1.execute _).expects(*, Phase.VERIFY).returning(TargetResult(target1, Phase.VERIFY, Status.SUCCESS, Instant.now()))
+
+        (targetP2.instantiate _).expects(*, *).once().returns(target2)
+        (target2.name _).expects().atLeastOnce().returns("target2")
+        (target2.identifier _).expects().atLeastOnce().returns(TargetIdentifier("target2"))
+        (target2.before _).expects().atLeastOnce().returns(Seq())
+        (target2.after _).expects().atLeastOnce().returns(Seq())
+        (target2.phases _).expects().atLeastOnce().returns(Set(Phase.VALIDATE, Phase.CREATE, Phase.BUILD))
+        (target2.metadata _).expects().returns(Metadata(name = "target2", kind = "target", category = "target"))
+        (target2.requires _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target2.provides _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target2.digest _).expects(Phase.VALIDATE).atLeastOnce().returns(TargetDigest("default", "project", "target2", Phase.VALIDATE))
+        (target2.dirty _).expects(*, Phase.VALIDATE).returns(Yes)
+        (target2.execute _).expects(*, Phase.VALIDATE) returning (TargetResult(target2, Phase.VALIDATE, Status.SUCCESS, Instant.now()))
+        (target2.requires _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target2.provides _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target2.requires _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+        (target2.provides _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+
+        (targetP3.instantiate _).expects(*, *).once().returns(target3)
+        (target3.identifier _).expects().atLeastOnce().returns(TargetIdentifier("target3"))
+        (target3.before _).expects().atLeastOnce().returns(Seq())
+        (target3.after _).expects().atLeastOnce().returns(Seq())
+        (target3.phases _).expects().atLeastOnce().returns(Set(Phase.CREATE))
+        (target3.requires _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target3.provides _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target3.requires _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target3.provides _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target3.requires _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+        (target3.provides _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+
+        val coord = new JobCoordinator(session)
+        coord.execute(job, Lifecycle.ofPhase(Phase.VERIFY)) should be(Status.SUCCESS)
+    }
+
+    it should "only execute selected targets" in {
+        val target1 = mock[Target]
+        val target2 = mock[Target]
+        val target3 = mock[Target]
+        val targetP1 = mock[Prototype[Target]]
+        val targetP2 = mock[Prototype[Target]]
+        val targetP3 = mock[Prototype[Target]]
+        val project = Project(
+            "project",
+            targets = Map(
+                "target1" -> targetP1,
+                "target2" -> targetP2,
+                "target3" -> targetP3
+            )
+        )
+        val session = Session.builder().withSparkSession(spark).build()
+        val context = session.getContext(project)
+        val job = Job(
+            Job.Properties(context),
+            targets = Seq(
+                TargetIdentifier("target1"),
+                TargetIdentifier("target2"),
+                TargetIdentifier("target3")
+            )
+        )
+
+        (targetP1.instantiate _).expects(*, *).once().returns(target1)
+        (target1.name _).expects().atLeastOnce().returns("target1")
+        (target1.identifier _).expects().atLeastOnce().returns(TargetIdentifier("target1"))
+        (target1.before _).expects().atLeastOnce().returns(Seq())
+        (target1.after _).expects().atLeastOnce().returns(Seq())
+        (target1.phases _).expects().atLeastOnce().returns(Set(Phase.VALIDATE, Phase.CREATE, Phase.VERIFY))
+        (target1.metadata _).expects().atLeastOnce().returns(Metadata(name = "target1", kind = "target", category = "target"))
+        (target1.requires _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target1.provides _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target1.digest _).expects(Phase.VALIDATE).atLeastOnce().returns(TargetDigest("default", "project", "target1", Phase.VALIDATE))
+        (target1.dirty _).expects(*, Phase.VALIDATE).returns(Yes)
+        (target1.execute _).expects(*, Phase.VALIDATE) returning (TargetResult(target1, Phase.VALIDATE, Status.SUCCESS, Instant.now()))
+        (target1.requires _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target1.provides _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target1.digest _).expects(Phase.CREATE).atLeastOnce().returns(TargetDigest("default", "project", "target1", Phase.BUILD))
+        (target1.dirty _).expects(*, Phase.CREATE).returns(Yes)
+        (target1.execute _).expects(*, Phase.CREATE).returning(TargetResult(target1, Phase.CREATE, Status.SUCCESS, Instant.now()))
+        (target1.requires _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+        (target1.provides _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+        (target1.digest _).expects(Phase.VERIFY).atLeastOnce().returns(TargetDigest("default", "project", "target1", Phase.VERIFY))
+        (target1.dirty _).expects(*, Phase.VERIFY).returns(Yes)
+        (target1.execute _).expects(*, Phase.VERIFY).returning(TargetResult(target1, Phase.VERIFY, Status.SUCCESS, Instant.now()))
+
+        (targetP2.instantiate _).expects(*, *).once().returns(target2)
+        (target2.identifier _).expects().atLeastOnce().returns(TargetIdentifier("target2"))
+        (target2.before _).expects().atLeastOnce().returns(Seq())
+        (target2.after _).expects().atLeastOnce().returns(Seq())
+        (target2.phases _).expects().atLeastOnce().returns(Set(Phase.VALIDATE, Phase.CREATE, Phase.BUILD))
+        (target2.requires _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target2.provides _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target2.requires _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target2.provides _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target2.requires _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+        (target2.provides _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+
+        (targetP3.instantiate _).expects(*, *).once().returns(target3)
+        (target3.identifier _).expects().atLeastOnce().returns(TargetIdentifier("target3"))
+        (target3.before _).expects().atLeastOnce().returns(Seq())
+        (target3.after _).expects().atLeastOnce().returns(Seq())
+        (target3.phases _).expects().atLeastOnce().returns(Set(Phase.CREATE))
+        (target3.requires _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target3.provides _).expects(Phase.VALIDATE).atLeastOnce().returns(Set())
+        (target3.requires _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target3.provides _).expects(Phase.CREATE).atLeastOnce().returns(Set())
+        (target3.requires _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+        (target3.provides _).expects(Phase.VERIFY).atLeastOnce().returns(Set())
+
+        val coord = new JobCoordinator(session)
+        coord.execute(job, Lifecycle.ofPhase(Phase.VERIFY), targets=Seq(".*1".r)) should be(Status.SUCCESS)
     }
 }